med_ai_simulator_telegram/ai_model/utils/requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
cycler==0.11.0
environs==8.0.0
fonttools==4.28.2
frozenlist==1.2.0
grad-cam==1.3.5
idna==3.3
imageio==2.13.1
joblib==1.1.0
kiwisolver==1.3.2
marshmallow==3.14.1
matplotlib==3.5.0
multidict==5.2.0
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
ttach==0.0.3
typing-extensions==4.0.1
yarl==1.7.2
